refactor(tasks): Optimizar payloads y buffers con float
fsmTask:
- receivedVertices[20] -> [10]: 320B -> 80B (ahorro 240B)
- sendPosition(): double -> float (payload 16B -> 8B)
- updatePosition(): validacion sizeof(float)*2
- Elimina updateFence() (limits creados en receivedFence)

loraTask:
- VERTEX_SIZE: 16B -> 8B
- Fragmentos: 2 vertices/msg -> 4 vertices/msg (mejora 100%)
- Validacion payload GPS: sizeof(float)

sensorAcqTask:
- MSG_ID_SEND_GPS payload: sizeof(float)*2

Impacto: -50% bytes en transmision, -40% mensajes LoRa

+++ CM4/Core/Src/threads/sensorAcqTask.cpp
@@ -88,7 +88,7 @@ void sensorAcqTask(void *argument) {
             switch (msgReceived->id) {
                 case MSG_ID_REQUEST_GPS:
                     msgToSend = MessagePool_Allocate();
-                    EmbeddedMessage_CreateWithPayload(msgToSend, MSG_ID_SEND_GPS, MODULE_SENSOR_ACQ, MODULE_FSM, (uint8_t*)gpsData, sizeof(double) * 2);
+                    EmbeddedMessage_CreateWithPayload(msgToSend, MSG_ID_SEND_GPS, MODULE_SENSOR_ACQ, MODULE_FSM, (uint8_t*)gpsData, sizeof(float) * 2);
                     osMessageQueuePut(dispatcherQueueHandle, &msgToSend, 0, 0);
                     RTOS_LOG_DEBUG("[SENSOR_ACQ]] Sent GPS data to FSM\n");
                     msgToSend = NULL;
